Added ajax slide.js will be trying some special effects with going to the next image using jquery later on.
Fixed the key.js navigation script to know when there is no previous strip. Still
need to add upper limit on this one.

Changed the template, for viewing to view.html instead of view_strip.html this is for testing.

Added a ajax_image call to webserver.py this is tied in with the ajax requests i want to use.

Added the update queries that update the view cache. This is still pretty dumb as in only updates the last
page viewd on a strip. (overwriting the previous page that was viewed. Will expadn this later to per strip)